🕊 move types to types folder
