+++ app/backend/core/types/BrainstormResult.py
