+++ app/backend/core/types/ChatResult.py
