+++ app/backend/core/types/SummarizeResult.py
