+++ app/backend/summarize/summarize.py
@@ -11,7 +11,7 @@ from core.datahelper import Repository, Requestinfo
 from core.textsplit import splitPDF, splitText
 from core.types.Config import ApproachConfig
 from core.types.LlmConfigs import LlmConfigs
-from summarize.summarizeresult import SummarizeResult
+from core.types.SummarizeResult import SummarizeResult
 
 
 class DenserSummary(BaseModel):
